Add files via upload
First commit. File descriptions and functionality are as follows:


TEST.PY (NOT FOR MASTER)
test.py is a file used for testing the timing of YFinance data requests. The file uses one stock, MSFT, and makes various requests for price data. These include a request for the maximum price history, 1 month, 1 day, and intraday 1 minute prices. Current time print statements are dispersed between these calls to allow us to track the time it takes via terminal prints. This is NOT development or master branch code, this is merely an exercise to ascertain the feasibility of making requests for price data in a timely fashion.

TEST2.py (NOT FOR MASTER)
test2.py had a similar use-case as test.py, but the primary purpose was to check how often the latest price in the YFinance "1d" dataframe was updated. This appears to be roughly once a minute, as later confirmed by pulling the intraday 1 minute data which has datapoints every minute at the (hour):(minute):05 mark.

TEST3.PY (NEEDS REFACTORING)
test3.py contains somewhat hard-coded backend logic for creating a single YFinance ticker object for Microsoft and then requesting various data for it. There are additional functions for stripping down the price dataframes returned by YFinance in order to get a series of just closing prices (rather than Open, High, Low, Close prices and Volume) and for retrieving the most recent price from a yFinance dataframe of any size. Finally, the code at the bottom of the file uses this data stripped from the YFinance dataframes to create a MatPlotLib chart of the closing prices for the last month. This chart is then saved in PNG format so that it can be used later on in the Kivy framework as an IMAGE object.
This code will eventually be refactored and made extensible so that the actions taken on Microsoft in this file can be easily taken on any stock symbol passed to the code. Future refactoring will also allow for the request of, and chart creation from, data on timescales other than 1 month.

TEST4.py (NOT FOR MASTER)
test4.py contains an initial attempt at a GUI, using only python code and Kivy basics rather than python code accompanied by a .kv file, which allows for more customization. Because of the limitations discovered early on, this approach was abandoned and the file ins included merely as a demonstration.

TESTCHART.PNG
testchart.png is the current static chart being used as the IMAGE by the Kivy FloatLayout. It is a 1 month chart of Microsoft closing prices, created and saved by test3.py and called by the testGUI.kv file. It will be written over every time the code in test3.py is run.

TESTGUI.KV (WORK IN PROGRESS)
testgui.kv is the kv file half of our future GUI. It uses KVLang to delineate the appearance and size of the various buttons and images that make up the GUI window. Python code can not be executed directly in a .kv file, but triggers on buttons and other user interactions can be used to execute code in the accompanying testGUIwithkv.py file.

TESTGUIWITHKV.PY
testGUIwithkv.py contains the initializing logic for the Kivy FloatLayout class along with some basic calls to the Kivy Window to set the size of the screen and certain rudimentary characteristics. The Launch class is where future code executed via triggers in the .kv file must be located. The lone function here now, testFunction, is included as a simple example. It is triggered by clicking the SELL button on the GUI and prints a basic confirmation message to the terminal. In the future, the timeframe buttons (1d, 6mo, max, etc) will trigger functions here that call on code currently contained in test3.py in order to request the relevant price data and build the chart.