User Stories 3, 4, and 32. Tasks 33, 42, and 43
Tasks: #33, #42, #43. US: # 3, #4, #32. This takes care of tasks 42 and 43 as well as User Stories 3 and 4. These create a pop up when shares are bought/sold.  The task 33 and User Story 32 displays the financials table on the right side of the application. The commit is all together because the pop up for buy/sell button is very similar. Also, for the financial table, there were only few changes necessary between Sprint 2 and Sprint 3 so there was not a separate commit for it. Fixing the commit error.